+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 0.4.0 - dev
 
 + Add option `-` to `mix recode` to read from stdin.
-+ Add `Recode.Task.UnusedVariable`
++ Add `Recode.Task.UnusedVariable`.
++ Update `Recode.Task.SinglePipe`. Some false positives are fixed.
++ Update `Recode.Task.PipeFunOne`. Some false positives are fixed.
 
 ## 0.3.0 - 2022/08/28